refector: Change package structure in test
