+++ src/test/java/com/clone/ohouse/store/unused/BedRepositoryTest.java
@@ -1,4 +1,5 @@
 //package com.clone.ohouse.shop.product.domain.access;
+
 //
 //
 //import com.clone.ohouse.shop.product.domain.entity.Bed;
